requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.2
aisuite[all]>=0.1.0
openai>=1.3.5
langchain==0.1.4
langchain-community==0.0.14
langchain-core==0.1.23
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pydantic==2.4.2
python-dotenv==1.0.0
unstructured[pdf]>=0.10.30
pymupdf>=1.23.8
pillow>=10.0.0
tesseract-ocr>=5.3.3 